src/requirements.txt
# Canonical dependency list for the entire src/ tree.
# Install with: pip install -r src/requirements.txt

# API/runtime
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0
python-dotenv>=1.0

# HTTP + LLM providers
httpx>=0.27
openai>=1.30
google-genai>=1.0.0

# Retrieval and embeddings
fastembed>=0.3.0
torch>=2.0
transformers>=4.0

# Data ingestion
pandas>=2.0
openpyxl>=3.1
